Move tokenizer into extension/llm/tokenizer
