+++ examples/models/llama2/CMakeLists.txt
@@ -97,8 +97,8 @@ if(EXECUTORCH_USE_TIKTOKEN)
   set(ABSL_PROPAGATE_CXX_STD ON)
   set(_pic_flag ${CMAKE_POSITION_INDEPENDENT_CODE})
   set(CMAKE_POSITION_INDEPENDENT_CODE ON)
-  add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/third-party/abseil-cpp)
-  add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/third-party/re2)
+  add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/../../../extension/llm/third-party/abseil-cpp)
+  add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/../../../extension/llm/third-party/re2)
   set(CMAKE_POSITION_INDEPENDENT_CODE ${_pic_flag})
   target_link_libraries(llama_runner PUBLIC re2::re2)
 endif()
