+++ examples/models/llama2/tokenizer/targets.bzl
@@ -1,44 +1,23 @@
 load("@fbsource//xplat/executorch/build:runtime_wrapper.bzl", "runtime")
 
 def define_common_targets():
-    runtime.cxx_library(
-        name = "bpe_tokenizer",
-        srcs = [
-            "bpe_tokenizer.cpp",
-        ],
-        exported_headers = [
-            "tokenizer.h",
-            "bpe_tokenizer.h",
-        ],
-        exported_deps = [
-            "//executorch/runtime/core/exec_aten:lib",
-            "//executorch/runtime/core/exec_aten/util:scalar_type_util",
-        ],
-        visibility = [
-            "@EXECUTORCH_CLIENTS",
-        ],
-    )
+    """Defines targets that should be shared between fbcode and xplat.
 
+    The directory containing this targets.bzl file should also contain both
+    TARGETS and BUCK files that call this function.
+    """
     runtime.cxx_library(
         name = "tiktoken",
         srcs = [
-            "tiktoken.cpp",
             "llama_tiktoken.cpp",
         ],
         exported_headers = [
-            "tokenizer.h",
-            "tiktoken.h",
             "llama_tiktoken.h",
-            "base64.h",
         ],
         exported_deps = [
-            "//executorch/runtime/core/exec_aten:lib",
-            "//executorch/runtime/core/exec_aten/util:scalar_type_util",
+            "//executorch/extension/llm/tokenizer:tiktoken",
         ],
         visibility = [
             "@EXECUTORCH_CLIENTS",
-        ],
-        exported_external_deps = [
-            "re2",
         ],
     )
