+++ examples/models/llama2/tokenizer/test/test_tiktoken.cpp
@@ -7,7 +7,6 @@
  */
 
 #include <executorch/examples/models/llama2/tokenizer/llama_tiktoken.h>
-#include <executorch/examples/models/llama2/tokenizer/tokenizer.h>
 #include <executorch/runtime/platform/runtime.h>
 #include <gtest/gtest.h>
 #include <vector>
@@ -16,19 +15,6 @@ using namespace ::testing;
 
 namespace torch {
 namespace executor {
-
-class TiktokenExtensionTest : public Test {
- public:
-  void SetUp() override {
-    torch::executor::runtime_init();
-    tokenizer_ = get_tiktoken_for_llama();
-    modelPath_ = std::getenv("RESOURCES_PATH") +
-        std::string("/test_tiktoken_tokenizer.model");
-  }
-
-  std::unique_ptr<Tokenizer> tokenizer_;
-  std::string modelPath_;
-};
 
 class MultimodalTiktokenV5ExtensionTest : public Test {
  public:
@@ -43,41 +29,12 @@ class MultimodalTiktokenV5ExtensionTest : public Test {
   std::string modelPath_;
 };
 
-TEST_F(TiktokenExtensionTest, EncodeWithoutLoadFails) {
-  Result<std::vector<uint64_t>> res = tokenizer_->encode("hello world", 0, 0);
-  EXPECT_EQ(res.error(), Error::NotSupported);
-}
-
-TEST_F(TiktokenExtensionTest, DecodeWithoutLoadFails) {
-  auto result = tokenizer_->decode(0, 0);
-  EXPECT_EQ(result.error(), Error::NotSupported);
-}
-
-TEST_F(TiktokenExtensionTest, TokenizerVocabSizeIsExpected) {
-  Error res = tokenizer_->load(modelPath_.c_str());
-  EXPECT_EQ(res, Error::Ok);
-  EXPECT_EQ(tokenizer_->vocab_size(), 128256);
-  EXPECT_EQ(tokenizer_->bos_tok(), 128000);
-  EXPECT_EQ(tokenizer_->eos_tok(), 128001);
-}
-
 TEST_F(MultimodalTiktokenV5ExtensionTest, TokenizerVocabSizeIsExpected) {
   Error res = tokenizer_->load(modelPath_.c_str());
   EXPECT_EQ(res, Error::Ok);
   EXPECT_EQ(tokenizer_->vocab_size(), 128256);
   EXPECT_EQ(tokenizer_->bos_tok(), 128000);
   EXPECT_EQ(tokenizer_->eos_tok(), 128001);
-}
-
-TEST_F(TiktokenExtensionTest, TokenizerEncodeCorrectly) {
-  Error res = tokenizer_->load(modelPath_.c_str());
-  EXPECT_EQ(res, Error::Ok);
-  Result<std::vector<uint64_t>> out = tokenizer_->encode("hello world", 1, 0);
-  EXPECT_EQ(out.error(), Error::Ok);
-  EXPECT_EQ(out.get().size(), 3);
-  EXPECT_EQ(out.get()[0], 128000);
-  EXPECT_EQ(out.get()[1], 15339);
-  EXPECT_EQ(out.get()[2], 1917);
 }
 
 TEST_F(MultimodalTiktokenV5ExtensionTest, TokenizerEncodeCorrectly) {
@@ -101,18 +58,6 @@ TEST_F(MultimodalTiktokenV5ExtensionTest, TokenizerEncodeCorrectly) {
   }
 }
 
-TEST_F(TiktokenExtensionTest, TokenizerDecodeCorrectly) {
-  Error res = tokenizer_->load(modelPath_.c_str());
-  EXPECT_EQ(res, Error::Ok);
-  std::vector<std::string> expected = {"<|begin_of_text|>", "hello", " world"};
-  std::vector<uint64_t> tokens = {128000, 15339, 1917};
-  for (size_t i = 0; i < tokens.size(); i++) {
-    Result<std::string> out = tokenizer_->decode(0, tokens[i]);
-    EXPECT_EQ(out.error(), Error::Ok);
-    EXPECT_EQ(out.get(), expected[i]);
-  }
-}
-
 TEST_F(MultimodalTiktokenV5ExtensionTest, TokenizerDecodeCorrectly) {
   Error res = tokenizer_->load(modelPath_.c_str());
   EXPECT_EQ(res, Error::Ok);
@@ -133,45 +78,6 @@ TEST_F(MultimodalTiktokenV5ExtensionTest, TokenizerDecodeCorrectly) {
     EXPECT_EQ(out.error(), Error::Ok);
     EXPECT_EQ(out.get(), expected[i]);
   }
-}
-
-TEST_F(TiktokenExtensionTest, TokenizerDecodeOutOfRangeFails) {
-  Error res = tokenizer_->load(modelPath_.c_str());
-  EXPECT_EQ(res, Error::Ok);
-  // The vocab size is 128256, addes 256 just so the token is out of vocab
-  // range.
-  Result<std::string> out = tokenizer_->decode(0, 128256 + 256);
-  EXPECT_EQ(out.error(), Error::NotSupported);
-}
-
-TEST_F(TiktokenExtensionTest, ConstructionWithInvalidBOSIndex) {
-  // gtest death test doesn't work on iOS:
-  // https://github.com/google/googletest/issues/2834
-#if !GTEST_OS_IOS
-  EXPECT_EXIT(
-      std::make_unique<Tiktoken>(
-          std::make_unique<std::vector<std::string>>(
-              std::vector<std::string>{"<|end_of_text|>"}),
-          1,
-          0),
-      ::testing::KilledBySignal(SIGABRT),
-      "");
-#endif
-}
-
-TEST_F(TiktokenExtensionTest, ConstructionWithInvalidEOSIndex) {
-  // gtest death test doesn't work on iOS:
-  // https://github.com/google/googletest/issues/2834
-#if !GTEST_OS_IOS
-  EXPECT_EXIT(
-      std::make_unique<Tiktoken>(
-          std::make_unique<std::vector<std::string>>(
-              std::vector<std::string>{"<|begin_of_text|>"}),
-          0,
-          1),
-      ::testing::KilledBySignal(SIGABRT),
-      "");
-#endif
 }
 } // namespace executor
 } // namespace torch
