+++ extension/llm/tokenizer/targets.bzl
@@ -36,3 +36,43 @@ def define_common_targets():
             ":tokenizer_py_lib",
         ],
     )
+
+    runtime.cxx_library(
+        name = "bpe_tokenizer",
+        srcs = [
+            "bpe_tokenizer.cpp",
+        ],
+        exported_headers = [
+            "tokenizer.h",
+            "bpe_tokenizer.h",
+        ],
+        exported_deps = [
+            "//executorch/runtime/core/exec_aten:lib",
+            "//executorch/runtime/core/exec_aten/util:scalar_type_util",
+        ],
+        visibility = [
+            "@EXECUTORCH_CLIENTS",
+        ],
+    )
+
+    runtime.cxx_library(
+        name = "tiktoken",
+        srcs = [
+            "tiktoken.cpp",
+        ],
+        exported_headers = [
+            "tokenizer.h",
+            "tiktoken.h",
+            "base64.h",
+        ],
+        exported_deps = [
+            "//executorch/runtime/core/exec_aten:lib",
+            "//executorch/runtime/core/exec_aten/util:scalar_type_util",
+        ],
+        visibility = [
+            "@EXECUTORCH_CLIENTS",
+        ],
+        exported_external_deps = [
+            "re2",
+        ],
+    )
