+++ extension/llm/tokenizer/test/CMakeLists.txt
@@ -17,15 +17,16 @@ project(tokenizer_test)
 # Use C++17 for test.
 set(CMAKE_CXX_STANDARD 17)
 
-set(EXECUTORCH_ROOT ${CMAKE_CURRENT_SOURCE_DIR}/../../../../..)
+set(EXECUTORCH_ROOT ${CMAKE_CURRENT_SOURCE_DIR}/../../../..)
 
 include(${EXECUTORCH_ROOT}/build/Test.cmake)
 
 set(
   _tokenizer_test_srcs
   test_tiktoken.cpp
-  ${CMAKE_CURRENT_SOURCE_DIR}/../../../../../extension/llm/tokenizer/tiktoken.cpp
-  ${CMAKE_CURRENT_SOURCE_DIR}/../llama_tiktoken.cpp
+  test_bpe_tokenizer.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/../tiktoken.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/../bpe_tokenizer.cpp
 )
 
 set(ENV{RESOURCES_PATH} ${CMAKE_CURRENT_SOURCE_DIR}/resources)
@@ -33,8 +34,8 @@ set(ABSL_ENABLE_INSTALL ON)
 set(ABSL_PROPAGATE_CXX_STD ON)
 set(_pic_flag ${CMAKE_POSITION_INDEPENDENT_CODE})
 set(CMAKE_POSITION_INDEPENDENT_CODE ON)
-add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/../../../../../extension/llm/third-party/abseil-cpp ${CMAKE_CURRENT_BINARY_DIR}/abseil-cpp)
-add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/../../../../../extension/llm/third-party/re2 ${CMAKE_CURRENT_BINARY_DIR}/re2)
+add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/../../third-party/abseil-cpp ${CMAKE_CURRENT_BINARY_DIR}/abseil-cpp)
+add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/../../third-party/re2 ${CMAKE_CURRENT_BINARY_DIR}/re2)
 set(CMAKE_POSITION_INDEPENDENT_CODE ${_pic_flag})
 
 et_cxx_test(
@@ -45,5 +46,5 @@ et_cxx_test(
   re2::re2
 )
 target_include_directories(
-  tokenizer_test PRIVATE ${CMAKE_INSTALL_PREFIX}/include ${CMAKE_CURRENT_SOURCE_DIR}/../../../../../extension/llm/third-party/abseil-cpp
+  tokenizer_test PRIVATE ${CMAKE_INSTALL_PREFIX}/include ${CMAKE_CURRENT_SOURCE_DIR}/../../third-party/abseil-cpp
 )
