+++ extension/llm/tokenizer/test/targets.bzl
@@ -16,3 +16,39 @@ def define_common_targets():
             "//executorch/extension/llm/tokenizer:tokenizer_py_lib",
         ],
     )
+
+    runtime.cxx_test(
+        name = "test_bpe_tokenizer",
+        srcs = [
+            "test_bpe_tokenizer.cpp",
+        ],
+        deps = [
+            "//executorch/extension/llm/tokenizer:bpe_tokenizer",
+        ],
+        env = {
+            "RESOURCES_PATH": "$(location :resources)/resources",
+        },
+    )
+
+    runtime.cxx_test(
+        name = "test_tiktoken",
+        srcs = [
+            "test_tiktoken.cpp",
+        ],
+        deps = [
+            "//executorch/extension/llm/tokenizer:tiktoken",
+        ],
+        env = {
+            "RESOURCES_PATH": "$(location :resources)/resources",
+        },
+        external_deps = [
+            "re2",
+        ],
+    )
+
+    runtime.filegroup(
+        name = "resources",
+        srcs = native.glob([
+            "resources/**",
+        ]),
+    )
