+++ extension/llm/tokenizer/tiktoken.cpp
@@ -25,8 +25,8 @@
    limitations under the License.
  *************************************************************************/
 
-#include <executorch/examples/models/llama2/tokenizer/base64.h>
-#include <executorch/examples/models/llama2/tokenizer/tiktoken.h>
+#include <executorch/extension/llm/tokenizer/base64.h>
+#include <executorch/extension/llm/tokenizer/tiktoken.h>
 #include <cctype>
 #include <cstdint>
 #include <fstream>
