+++ extension/llm/tokenizer/tiktoken.h
@@ -8,7 +8,7 @@
 
 #pragma once
 
-#include <executorch/examples/models/llama2/tokenizer/tokenizer.h>
+#include <executorch/extension/llm/tokenizer/tokenizer.h>
 #include <re2/re2.h>
 #include <cstdint>
 #include <functional>
