+++ test/run_oss_cpp_tests.sh
@@ -87,8 +87,10 @@ build_and_run_test() {
     -Bcmake-out/"${test_dir}"
   cmake --build cmake-out/"${test_dir}" -j9
 
-  if [[ "$test_dir" =~ .*tokenizer.* ]]; then
+  if [[ "$test_dir" =~ .*examples/models/llama2/tokenizer.* ]]; then
     RESOURCES_PATH=$(realpath examples/models/llama2/tokenizer/test/resources)
+  elif [[ "$test_dir" =~ .*extension/llm/tokenizer.* ]]; then
+    RESOURCES_PATH=$(realpath extension/llm/tokenizer/test/resources)
   else
     RESOURCES_PATH=$(realpath extension/module/test/resources)
   fi
